requirements.txt
pandas
meteostat
pyarrow
tqdm
numpy
scikit-learn
lightgbm
joblib
kagglehub
holidays
seaborn
matplotlib
python-dotenv
xgboost
catboost
langchain 
langchain-openai
pydantic==2.12.5
pydantic-settings==2.10.1
langgraph
deepeval==3.7.8
